+++ packages/emails/src/templates/register.tsx
@@ -1,7 +1,6 @@
 import { Section } from "@react-email/section";
 import { Trans } from "react-i18next/TransWithoutContext";
 
-import { EmailContext } from "../components/email-context";
 import { EmailLayout } from "../components/email-layout";
 import {
   Card,
@@ -10,6 +9,7 @@ import {
   Text,
   trackingWide,
 } from "../components/styled-components";
+import type { EmailContext } from "../types";
 
 interface RegisterEmailProps {
   code: string;
